🐬 Use Flipper for Feature Flagging (#5786)
* Add Flipper/Flipper-ActiveRecord

* Generate/Run Flipper migrations

* Configure Flipper

* Add `FeatureFlag` for managing feature flags

* Introduce FeatureFlagHelpers for tests

These are meant to make it more convenient to handle feature flags in
our tests. They may not be necessary, but my hope is that they will make
our lives eaiser.

* Add tests for feature flagging organizations

* Add FeatureFlag integration test

This doesn't really fit with the other integration tests as it's not
testing controller actions, but it didn't seem like just a model test,
so I put it in test/integrations.

* Add `flipper-ui` to Gemfile

* Pin gem versions

* Introduce FlipperActor to handle users/orgs

* Configure Flipper UI

* Make actor rendering a little friendlier

* Improve actor-management ergonomics

Flipper's default of `User;#{id}` isn't the easiest to type, nor is it
the most memorable, so I'm working to make `user:handle` the default.

* Improve the Flipper UI experience

* Add system tests for feature flagging

* Make placeholder text self-documenting

* Pacify Rubocop

* Use `at:` in routes for consistency

* Rename FlipperActor -> FeatureFlagActor

Per code reveiw feedback, FlipperActor was leaking an implementation
detail which we're better off containing.