Extend wait for WebAuthnVerificationTest
This system test has been causing a lot of builds to fail lately. I
believe the default wait time of two seconds just isn't enough for some
of the tests to pass. Rather than use an extended wait time across the
entire test suite, I'm introducing a module which can be mixed in to
take effect on a file-by-file basis.