⛳ Feature Flag Organizations (#5804)
* Fix typo in test helper

`#disable` and `#disable_globally` are synonyms, but I opted to only use
the latter. I missed this reference, however. Fixing it now.

* Fix show/edit confusion

The route for this maps `confirm` to `edit` rather than `show`

* Gate organizations functionality

As we roll out the organizations beta, we want to ensure we've got
fine-grained control over enrollement.

* Use a constant to prevent typos

Symbols are too easy to mess up.

* Permit broad access to organizations show/index

Per code review feedback, we don't need to gate access to the show and
index pages. They should be public.

* Feature flag invitation logic

This page is meant to be public, but organizations are a beta feature.
We need to gate the logic so that we don't expose the feature too
broadly.

* Only introduce organizations when ff is enabled