😶 Fix Trusted Publisher Creation with Blank Environment (#5819)
* Introduce a failing test

This behavior was reported recently, but it takes a special set of
interactions to reproduce. This is meant to serve as a starting point
for a fix.

* Remove invalid validation

We cannot require presence and allow nil at the same time. Since we _do_
allow blank values, I'm keeping the length validation.

* Permit intentionally blank "environment" values

This came up when a team had created a trusted publisher with values
which they no longer wished to use. When they attempted to create a new
trusted publisher (with no environment value), it fetched the previously
created trusted publisher record which was not destroyed.

* Coerce blank values to `nil`

We were getting this behavior from the stack (per local testing), but
we're better off being explicit.