Make ownership an organization OR user concern
Until now, we've allowed a `Rubygem` to be owned by users (via
`Ownership`s) and an `Organization`. This has some downsides. For
example, we've had to explicitly check both ownerships and organization
membership in our gem-related policies.

I've opted to push this logic into `Rubygem` and enforce the ownership
check to be organization-specific if the gem belongs to an organization
OR ownership-specific (i.e. user-specific) if it is not.