Prevent org handles from clashing with routes
In order to view an organization's dashboard, a user needs to visit
`/organizations/{handle}`. However, we use `/organizations` for other
routing as well. For example, `/organizations/onboarding` is used to
bring new organizations onto the application.

In order to prevent confusion, we should reserve specific patterns so
they cannot be used.

**NOTE:** This is not meant to be comprehensive, but a starting point
from which we can build. I believe we're better off establishing a
pattern than getting it perfect out of the gate.