🔁 Allow Membership Invitations to Be Resent (#5873)
* Allow membership invitations to be resent

When a member has been invited, but has not accepted their invitation,
we want to allow an admin+ to send another invitation. I chose to place
the button for this interaction in the same place as the "joined" date
when the invitation has been confirmed.

* Ensure non-English locales are present

* Use `ButtonComponent` instead of a full form

* Move "Resend Invitation" button next to "Update"

* Use `type: :link` to use a self-containeed form

We've got some surprising behavior in ButtonComponent, which we'll have
to resolve at some point. For the moment, I'll use `type: :link` to get
the behavior and appearance that I want.