🙅 Prevent Org Handles from Clashing with Routes  (#5882)
* Prevent org handles from clashing with routes

In order to view an organization's dashboard, a user needs to visit
`/organizations/{handle}`. However, we use `/organizations` for other
routing as well. For example, `/organizations/onboarding` is used to
bring new organizations onto the application.

In order to prevent confusion, we should reserve specific patterns so
they cannot be used.

**NOTE:** This is not meant to be comprehensive, but a starting point
from which we can build. I believe we're better off establishing a
pattern than getting it perfect out of the gate.

* Allow `rubygems` because it's in the seeds file

This highlights that we'll need to manually update this list while we
figure out the workflow. We may want to allow some of these names, but
IMO it's better to be overly restrictive for now.

* Pare down the list of reserved names

Per code review feedback, we can be less restrictive and revisit this
once organizations are more adopted.