Fix CodeQL Parser Errors (#5926)
* Help CodeQL parse concerns

CodeQL has been failing to evaluate these files because it believes
(erroneously) that this is invalid syntax. I've opened an issue with the
CodeQL to have this fixed, but we may as well benefit from the work that
it _could_ do if we had the "right" syntax.

* Relax block forwarding rules

This allows us to be compliant with Rubocop _and_ CodeQL. Rubocop is
actually allowing us to use some pretty varying approaches, so loosening
restrictions just a bit further.