Introduce `.editorconfig` for consistency
I've seen a couple requests lately to add newlines at the end of a file,
both by humans and coding agents. These standards are enforced by
Rubocop, and should be picked up by various tools. They're not meant to
be new rules, but hints for tooling to make a contributor' (or
reviewer's) job easier.