Specifically call the "verify" method on Gem packages
I would like to purposely call the "verify" method on Gem::Package.

A side-effect of calling the `spec` method is that the `verify` method
gets called:

  https://github.com/rubygems/rubygems/blob/1cfe20f1fb9b74c2254c22e0bfe9cb4d551d42d7/lib/rubygems/package.rb#L602-L606

But I think RubyGems.org should be _intentionally_ verifying the package
contents (not just as a side-effect).  The functionality should be the
same, but I think this change better communicates the intent of the
code.