Prepare for (graceful) secret_key_base rotation (#6018)
* Prepare for (graceful) secret_key_base rotation

By extending the existing cookie_rotator.rb (used to migrate to
SHA-256), we can achieve just-in-time rotation for sessions. This should
keep users from being logged out.

I've introduced `PREVIOUS_SECRET_KEY_BASE`, and set it to the current
value of secret_key_base so that we can make a minimal change when
rotating. This _should_ allow us to rotate when needed going forward as
well.

* Use Rails's current key_generator

I made a mistake in using SHA1. This should have been SHA256, which is
what `key_generator` uses.

* Add PREVIOUS_SECRET_KEY_BASE to deploys

---------

Co-authored-by: Colby Swandale <996377+colby-swandale@users.noreply.github.com>