Introduce maintenance task to unconfirmed users
Sometimes accounts are created by someone other than the owner of an
email address. In these cases, a User is created, but the account is
never confirmed. If someone has been in this scenario for more than a
month, we should remove the record in order to ensure the owner of the
email is not contacted.